requirements.txt
websockets==12.0
prettytable==3.9.0
PyYAML==6.0.1 